convert data-feed logic to cereal tasks
