+++ components/data-feed-service/service/data-feed-service.go
@@ -5,20 +5,15 @@ import (
 	"compress/gzip"
 	"context"
 	"encoding/json"
-	"fmt"
 	"net/http"
-	"strings"
 	"time"
 
-	"github.com/golang/protobuf/ptypes"
 	"github.com/pkg/errors"
 	log "github.com/sirupsen/logrus"
 	rrule "github.com/teambition/rrule-go"
 
 	secrets "github.com/chef/automate/api/external/secrets"
-	cfgmgmtRequest "github.com/chef/automate/api/interservice/cfgmgmt/request"
 	cfgmgmtResponse "github.com/chef/automate/api/interservice/cfgmgmt/response"
-	cfgmgmt "github.com/chef/automate/api/interservice/cfgmgmt/service"
 	"github.com/chef/automate/components/compliance-service/api/reporting"
 	"github.com/chef/automate/components/data-feed-service/config"
 	"github.com/chef/automate/components/data-feed-service/dao"
@@ -27,12 +22,6 @@ import (
 
 	//"github.com/chef/automate/lib/cereal/patterns"
 	"github.com/chef/automate/lib/grpc/secureconn"
-)
-
-var (
-	dataFeedTaskName     = cereal.NewTaskName("data-feed-poll")
-	dataFeedWorkflowName = cereal.NewWorkflowName("data-feed-workflow")
-	dataFeedScheduleName = "periodic-data-feed-workflow"
 )
 
 type datafeedNotification struct {
@@ -80,11 +69,40 @@ func Start(dataFeedConfig *config.DataFeedConfig, connFactory *secureconn.Factor
 		return err
 	}
 
+	dataFeedClientTask, err := NewDataFeedClientTask(dataFeedConfig, connFactory, db, manager)
+	if err != nil {
+		return err
+	}
+
+	dataFeedComplianceTask, err := NewDataFeedComplianceTask(dataFeedConfig, connFactory, db, manager)
+	if err != nil {
+		return err
+	}
+	dataFeedNotifierTask, err := NewDataFeedNotifierTask(dataFeedConfig, connFactory, db, manager)
+	if err != nil {
+		return err
+	}
+
 	err = manager.RegisterWorkflowExecutor(dataFeedWorkflowName, &DataFeedWorkflowExecutor{workflowName: dataFeedWorkflowName})
 	if err != nil {
 		return err
 	}
-	err = manager.RegisterTaskExecutor(dataFeedTaskName, dataFeedPollTask, cereal.TaskExecutorOpts{
+	err = manager.RegisterTaskExecutor(dataFeedPollTaskName, dataFeedPollTask, cereal.TaskExecutorOpts{
+		Workers: 1,
+	})
+	if err != nil {
+		return err
+	}
+	err = manager.RegisterTaskExecutor(dataFeedClientTaskName, dataFeedClientTask, cereal.TaskExecutorOpts{
+		Workers: 1,
+	})
+	if err != nil {
+		return err
+	}
+	err = manager.RegisterTaskExecutor(dataFeedComplianceTaskName, dataFeedComplianceTask, cereal.TaskExecutorOpts{
+		Workers: 1,
+	})
+	err = manager.RegisterTaskExecutor(dataFeedNotifierTaskName, dataFeedNotifierTask, cereal.TaskExecutorOpts{
 		Workers: 1,
 	})
 	if err != nil {
@@ -108,8 +126,10 @@ func Start(dataFeedConfig *config.DataFeedConfig, connFactory *secureconn.Factor
 		NextFeedStart:   zeroTime,
 		NextFeedEnd:     zeroTime,
 	}
-
-	dataFeedWorkflowParams := DataFeedWorkflowParams{PollTaskParams: dataFeedTaskParams}
+	dataFeedClientTaskParams := DataFeedClientTaskParams{NodeIDs: []string{}}
+	dataFeedComplianceTaskParams := DataFeedComplianceTaskParams{}
+	log.Infof("dataFeedComplianceTaskParams %v", dataFeedComplianceTaskParams)
+	dataFeedWorkflowParams := DataFeedWorkflowParams{PollTaskParams: dataFeedTaskParams, ClientTaskParams: dataFeedClientTaskParams, ComplianceTaskParams: dataFeedComplianceTaskParams}
 
 	err = manager.CreateWorkflowSchedule(context.Background(),
 		dataFeedScheduleName, dataFeedWorkflowName,
@@ -130,221 +150,9 @@ func Start(dataFeedConfig *config.DataFeedConfig, connFactory *secureconn.Factor
 	return manager.Start(context.Background())
 }
 
-type DataFeedWorkflowExecutor struct {
-	workflowName cereal.WorkflowName
-}
-
-type DataFeedWorkflowParams struct {
-	PollTaskParams DataFeedPollTaskParams
-}
-
-type DataFeedPollTask struct {
-	cfgMgmt   cfgmgmt.CfgMgmtClient
-	secrets   secrets.SecretsServiceClient
-	reporting reporting.ReportingServiceClient
-	db        *dao.DB
-	manager   *cereal.Manager
-}
-
-type DataFeedPollTaskParams struct {
-	AssetPageSize   int32
-	ReportsPageSize int32
-	FeedInterval    time.Duration
-	NextFeedStart   time.Time
-	NextFeedEnd     time.Time
-}
-
-type DataFeedWorkflowPayload struct {
-	Done bool
-}
-
-func (e *DataFeedWorkflowExecutor) OnStart(w cereal.WorkflowInstance, ev cereal.StartEvent) cereal.Decision {
-	params := DataFeedWorkflowParams{}
-	log.Infof("OnStart params %v", params)
-	err := w.GetParameters(&params)
-	if err != nil {
-		return w.Fail(err)
-	}
-
-	err = w.EnqueueTask(dataFeedTaskName, params)
-	if err != nil {
-		return w.Fail(err)
-	}
-
-	initialPayload := DataFeedWorkflowPayload{
-		Done: true,
-	}
-	return w.Continue(&initialPayload)
-}
-
-func (e *DataFeedWorkflowExecutor) OnTaskComplete(w cereal.WorkflowInstance, ev cereal.TaskCompleteEvent) cereal.Decision {
-	params := DataFeedWorkflowParams{}
-	err := w.GetParameters(&params)
-	if err != nil {
-		return w.Fail(err)
-	}
-
-	payload := DataFeedWorkflowPayload{}
-	err = w.GetPayload(&payload)
-	if err != nil {
-		return w.Fail(err)
-	}
-
-	err = ev.Result.Err()
-	if err != nil {
-		return w.Fail(err)
-	}
-
-	taskParams := DataFeedWorkflowParams{}
-	if err = ev.Result.GetParameters(&taskParams); err != nil {
-		return w.Fail(err)
-	}
-
-	if payload.Done {
-		return w.Complete()
-	}
-	return w.Continue(&payload)
-
-}
-
-func (e *DataFeedWorkflowExecutor) OnCancel(w cereal.WorkflowInstance, ev cereal.CancelEvent) cereal.Decision {
-	return w.Complete()
-}
-
-func NewDataFeedPollTask(dataFeedConfig *config.DataFeedConfig, connFactory *secureconn.Factory, db *dao.DB, manager *cereal.Manager) (*DataFeedPollTask, error) {
-
-	cfgMgmtConn, err := connFactory.Dial("config-mgmt-service", dataFeedConfig.CfgmgmtConfig.Target)
-	if err != nil {
-		return nil, errors.Wrap(err, "could not connect to config-mgmt-service")
-	}
-
-	secretsConn, err := connFactory.Dial("secrets-service", dataFeedConfig.SecretsConfig.Target)
-	if err != nil {
-		return nil, errors.Wrap(err, "could not connect to secrets-service")
-	}
-
-	complianceConn, err := connFactory.Dial("compliance-service", dataFeedConfig.ComplianceConfig.Target)
-	if err != nil {
-		return nil, errors.Wrap(err, "could not connect to compliance-service")
-	}
-
-	return &DataFeedPollTask{
-		secrets:   secrets.NewSecretsServiceClient(secretsConn),
-		reporting: reporting.NewReportingServiceClient(complianceConn),
-		cfgMgmt:   cfgmgmt.NewCfgMgmtClient(cfgMgmtConn),
-		db:        db,
-		manager:   manager,
-	}, nil
-}
-
-func (d *DataFeedPollTask) Run(ctx context.Context, task cereal.Task) (interface{}, error) {
-	params := DataFeedWorkflowParams{}
-	err := task.GetParameters(&params)
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to parse task parameters")
-	}
-	log.Infof("Run params %v", params)
-	now := time.Now()
-	params = getFeedTimes(params, now)
-	feedStartTime := params.PollTaskParams.NextFeedStart
-	feedEndTime := params.PollTaskParams.NextFeedEnd
-	destinations, err := d.getDestinations()
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to get destinations from db")
-	}
-	params.PollTaskParams.NextFeedStart = feedEndTime
-	params.PollTaskParams.NextFeedEnd = feedEndTime.Add(params.PollTaskParams.FeedInterval)
-	log.Debugf("Updated Feed interval start, end: %s, %s", params.PollTaskParams.NextFeedStart.Format("15:04:05"), params.PollTaskParams.NextFeedEnd.Format("15:04:05"))
-	err = d.manager.UpdateWorkflowScheduleByName(context.Background(), dataFeedScheduleName, dataFeedWorkflowName,
-		cereal.UpdateParameters(params),
-		cereal.UpdateEnabled(true))
-	if err != nil {
-		return nil, err
-	}
-
-	if len(destinations) == 0 {
-		return nil, nil
-	}
-	// build messages for any nodes which have had a CCR in last window, include last compliance report
-	datafeedMessages := d.buildDatafeed(ctx, params.PollTaskParams.AssetPageSize, feedStartTime, feedEndTime)
-	// build messages for any reports in last window include last CCR and OHAI
-	d.buildReportFeed(ctx, params.PollTaskParams.ReportsPageSize, feedStartTime, feedEndTime, datafeedMessages)
-	// get the valus from this ipaddress:datafeedMessage map
-	data := make([]datafeedMessage, 0, len(datafeedMessages))
-	for _, value := range datafeedMessages {
-		data = append(data, value)
-	}
-
-	if len(datafeedMessages) == 0 {
-		return nil, nil
-	}
-	for destination := range destinations {
-		log.Debugf("Destination name %v", destinations[destination].Name)
-		log.Debugf("Destination url %v", destinations[destination].URL)
-		log.Debugf("Destination secret %v", destinations[destination].Secret)
-
-		username, password, err := GetCredentials(ctx, d.secrets, destinations[destination].Secret)
-
-		if err != nil {
-			log.Errorf("Error retrieving credentials, cannot send asset notification: %v", err)
-			// TODO error handling - need some form of report in automate that indicates when data was sent and if it was successful
-		} else {
-			// build and send notification for this rule
-			notification := datafeedNotification{username: username, password: password, url: destinations[destination].URL, data: data}
-
-			client := NewDataClient()
-			err = send(client, notification)
-			if err != nil {
-				handleSendErr(notification, feedStartTime, feedEndTime, err)
-			}
-		}
-	}
-
-	return nil, nil
-}
-
-func (d *DataFeedPollTask) getDestinations() ([]dao.Destination, error) {
-	return d.db.ListDBDestinations()
-}
-
 func handleSendErr(notification datafeedNotification, startTime time.Time, endTime time.Time, err error) {
 	log.Errorf("Failed to send notification to %v. Start: %v, End: %v. %v", notification.url, startTime, endTime, err)
 	// TODO report this failure to the UI with the time window and notification details
-}
-
-func getFeedTimes(params DataFeedWorkflowParams, now time.Time) DataFeedWorkflowParams {
-
-	if params.PollTaskParams.NextFeedStart.IsZero() {
-		params.PollTaskParams.NextFeedEnd = getFeedEndTime(params.PollTaskParams.FeedInterval, now)
-		params.PollTaskParams.NextFeedStart = params.PollTaskParams.NextFeedEnd.Add(-params.PollTaskParams.FeedInterval)
-	}
-
-	log.Debugf("Current Feed interval start, end: %s, %s", params.PollTaskParams.NextFeedStart.Format("15:04:05"), params.PollTaskParams.NextFeedEnd.Format("15:04:05"))
-	return params
-}
-
-func getFeedEndTime(feedInterval time.Duration, now time.Time) time.Time {
-	log.Debugf("Time Now: %s", now.Format("15:04:05"))
-	/*
-	 * We round the current time down based on the interval duration to get the end of the last interval.
-	 *
-	 * Round will round either up or down to the nearest value of the inteval duration.
-	 * e.g 1:20pm rounds to 1pm, 1:40pm rounds to 2pm
-	 *
-	 * If we have rounded down that will be our feed end time. The end of a clock interval
-	 * rather than current time e.g. 1pm
-	 */
-	feedEndTime := now.Round(feedInterval)
-	log.Debugf("FeedInterval/Units: %s", feedInterval)
-	/*
-	 * If we have rounded up we subtract the interval to effectively round down
-	 */
-	if feedEndTime.After(now) {
-		feedEndTime = feedEndTime.Add(-feedInterval)
-		log.Debugf("feedEndTime after: %s", feedEndTime.Format("15:04:05"))
-	}
-	log.Debugf("feedEndTime: %s", feedEndTime.Format("15:04:05"))
-	return feedEndTime
 }
 
 func GetCredentials(ctx context.Context, client secrets.SecretsServiceClient, secretID string) (string, string, error) {
@@ -417,256 +225,4 @@ func (client DataClient) sendNotification(notification datafeedNotification) err
 		log.Warnf("Error clsoing response body %v", err)
 	}
 	return nil
-}
-
-func (d *DataFeedPollTask) buildDatafeed(ctx context.Context, pageSize int32, feedStartTime time.Time, feedEndTime time.Time) map[string]datafeedMessage {
-	log.Debug("Inventory nodes start")
-	feedStartString, err := ptypes.TimestampProto(feedStartTime)
-	if err != nil {
-		return nil
-	}
-	feedEndString, err := ptypes.TimestampProto(feedEndTime)
-	if err != nil {
-		return nil
-	}
-
-	nodesRequest := &cfgmgmtRequest.InventoryNodes{
-		PageSize: pageSize,
-		Start:    feedStartString,
-		End:      feedEndString,
-		Sorting: &cfgmgmtRequest.Sorting{
-			Order: cfgmgmtRequest.Order_desc,
-		},
-	}
-
-	inventoryNodes, err := d.cfgMgmt.GetInventoryNodes(ctx, nodesRequest)
-	if err != nil {
-		return nil
-	}
-
-	nodeIDs := make([]string, 0)
-	log.Debugf("No of inventory nodes %v", len(inventoryNodes.Nodes))
-	for len(inventoryNodes.Nodes) > 0 {
-		for _, node := range inventoryNodes.Nodes {
-			log.Debugf("Inventory node #%v", node)
-			nodeIDs = append(nodeIDs, node.Id)
-		}
-		lastNode := inventoryNodes.Nodes[len(inventoryNodes.Nodes)-1]
-		nodesRequest.CursorId = lastNode.Id
-		nodesRequest.CursorDate = lastNode.Checkin
-
-		inventoryNodes, err = d.cfgMgmt.GetInventoryNodes(ctx, nodesRequest)
-		log.Debugf("inventory nodes %v, cursor %v", len(inventoryNodes.Nodes), lastNode.Id)
-		if err != nil {
-			return nil
-		}
-	}
-	log.Info("Building data feed...")
-
-	nodeMessages := make(map[string]datafeedMessage)
-
-	for _, nodeID := range nodeIDs {
-		filters := []string{"id:" + nodeID}
-
-		message := getNodeData(ctx, d.cfgMgmt, filters)
-
-		ipaddress := message.Automatic["ipaddress"].(string)
-
-		filter := &reporting.ListFilter{Type: "ipaddress", Values: []string{ipaddress}}
-		reportFilters := []*reporting.ListFilter{filter}
-
-		ipaddressQuery := &reporting.Query{
-			Page:    0,
-			PerPage: 1,
-			Filters: reportFilters,
-			Sort:    "latest_report.end_time",
-			Order:   reporting.Query_DESC,
-		}
-		reports, err := d.reporting.ListReports(ctx, ipaddressQuery)
-		if err != nil {
-			log.Errorf("Error getting report by ip %v", err)
-			//TODO
-		}
-		// might not be a report!
-
-		//var fullReport *reporting.
-		fullReport := new(reporting.Report)
-		if len(reports.Reports) != 0 {
-			reportID := &reporting.Query{Id: reports.Reports[0].Id}
-			fullReport, err = d.reporting.ReadReport(ctx, reportID)
-			// TODO err handling
-			if err != nil {
-				log.Errorf("Error getting report by ip %v", err)
-				//TODO
-			}
-		}
-
-		message.Report = fullReport
-		nodeMessages[ipaddress] = message
-
-	}
-	log.Debugf("%v node attribute messages retrieved in interval", len(nodeMessages))
-	return nodeMessages
-}
-
-func getNodeData(ctx context.Context, client cfgmgmt.CfgMgmtClient, filters []string) datafeedMessage {
-
-	nodeFilters := &cfgmgmtRequest.Nodes{Filter: filters}
-	nodes, err := client.GetNodes(ctx, nodeFilters)
-	if err != nil {
-		log.Errorf("Error getting cfgmgmt/nodes %v", err)
-		//TODO
-	}
-
-	nodeStruct := nodes.Values[0].GetStructValue()
-	id := nodeStruct.Fields["id"].GetStringValue()
-	lastRunId := nodeStruct.Fields["latest_run_id"].GetStringValue()
-	nodeAttributes, err := client.GetAttributes(ctx, &cfgmgmtRequest.Node{NodeId: id})
-	if err != nil {
-		log.Errorf("Error getting attributes %v", err)
-		//TODO
-	}
-	var automaticJson map[string]interface{}
-	err = json.Unmarshal([]byte(nodeAttributes.Automatic), &automaticJson)
-	if err != nil {
-		log.Errorf("Could not parse automatic attributes from json: %v", err)
-		// continue to the next node
-		// TODO
-	}
-	attributesJson := buildDynamicJson(automaticJson)
-
-	lastRun, error := client.GetNodeRun(ctx, &cfgmgmtRequest.NodeRun{NodeId: id, RunId: lastRunId})
-
-	if error != nil {
-		log.Errorf("Error getting node run %v", error)
-	} else {
-		log.Debugf("Last run\n %v", lastRun)
-	}
-
-	return datafeedMessage{Automatic: attributesJson, LastRun: lastRun}
-}
-
-func buildDynamicJson(automaticJson map[string]interface{}) map[string]interface{} {
-	// check what format he json is e.g. ubuntu/ windows
-	// remove specific sections
-	// insert normalised
-	var dynamicJson map[string]interface{}
-	if automaticJson["lsb"] != nil {
-		dynamicJson = buildUbuntuJson(automaticJson)
-	} else if automaticJson["os"] == "windows" {
-		dynamicJson = buildWindowsJson(automaticJson)
-	}
-
-	return dynamicJson
-}
-
-func buildUbuntuJson(ubuntuJson map[string]interface{}) map[string]interface{} {
-	// check what format he json is e.g. ubuntu/ windows
-	// remove specific sections
-	// insert normalised
-
-	lsb := ubuntuJson["lsb"].(map[string]interface{})
-	ubuntuJson["description"] = lsb["description"]
-
-	dmi := ubuntuJson["dmi"].(map[string]interface{})
-	system := dmi["system"].(map[string]interface{})
-	ubuntuJson["serial_number"] = system["serial_number"]
-
-	//delete(ubuntuJson, "lsb")
-	//delete(ubuntuJson, "system")
-
-	return ubuntuJson
-}
-
-func buildWindowsJson(windowsJson map[string]interface{}) map[string]interface{} {
-	// check what format he json is e.g. ubuntu/ windows
-	// remove specific sections
-	// insert normalised
-
-	kernel := windowsJson["kernel"].(map[string]interface{})
-	windowsJson["description"] = kernel["name"]
-	osInfo := kernel["os_info"].(map[string]interface{})
-	windowsJson["serial_number"] = osInfo["serial_number"]
-	servicePackMajorVersion := fmt.Sprintf("%g", osInfo["service_pack_major_version"].(float64))
-	servicePackMinorVersion := fmt.Sprintf("%g", osInfo["service_pack_minor_version"].(float64))
-	servicePack := strings.Join([]string{servicePackMajorVersion, servicePackMinorVersion}, ".")
-	windowsJson["os_service_pack"] = servicePack
-
-	//delete(windowsJson, "kernel")
-	// TODO we could get the chassis from the kernel attributes
-
-	return windowsJson
-}
-
-func (d *DataFeedPollTask) buildReportFeed(ctx context.Context, pageSize int32, feedStartTime time.Time, feedEndTime time.Time, datafeedMessages map[string]datafeedMessage) {
-	feedStartString := strings.SplitAfter(feedStartTime.Format(time.RFC3339), "Z")[0]
-	feedEndString := strings.SplitAfter(feedEndTime.Format(time.RFC3339), "Z")[0]
-	log.Infof("Building report feed... %v - %v", feedStartString, feedEndString)
-
-	startFilter := &reporting.ListFilter{Type: "start_time", Values: []string{feedStartString}}
-	endFilter := &reporting.ListFilter{Type: "end_time", Values: []string{feedEndString}}
-
-	filters := []*reporting.ListFilter{startFilter, endFilter}
-
-	// page is not something we can configure we should start with
-	// page at 1 and work out how many calls to make based on page
-	// size divide the total by page size and add 1 and we loop
-	// over that
-	page := int32(1)
-	query := &reporting.Query{
-		Page:    page,
-		PerPage: pageSize,
-		Filters: filters,
-		Sort:    "latest_report.end_time",
-		Order:   reporting.Query_DESC,
-	}
-	log.Debugf("report query %v", query)
-
-	reports, err := d.reporting.ListReports(ctx, query)
-	if err != nil {
-		log.Errorf("Error getting reporting/ListReports %v", err)
-	}
-
-	pages := (reports.Total / pageSize)
-	if (reports.Total % pageSize) != 0 {
-		pages++
-	}
-	log.Debugf("Total reports: %v, reports per page: %v, total pages %v: ",
-		reports.Total, pageSize, pages)
-
-	// get reports from the pages
-	for page <= pages {
-		log.Debugf("report query %v", query)
-		for report := range reports.Reports {
-			ipaddress := reports.Reports[report].Ipaddress
-			log.Debugf("report has ipaddress %v", ipaddress)
-			if _, ok := datafeedMessages[ipaddress]; ok {
-				log.Debugf("node data already exists for %v", ipaddress)
-			} else {
-				// the node hasn't had a client run in last window, but has a report so we can add the report
-				id := &reporting.Query{Id: reports.Reports[report].Id}
-
-				fullReport, err := d.reporting.ReadReport(ctx, id)
-				if err != nil {
-					log.Debugf("Error getting report by if %v", err)
-					//TODO
-				}
-				filters := []string{"ipaddress:" + ipaddress}
-				// node the latest node data associated with this report
-				message := getNodeData(ctx, d.cfgMgmt, filters)
-				message.Report = fullReport
-				datafeedMessages[ipaddress] = message
-			}
-
-		}
-		page++
-		query = &reporting.Query{
-			Page:    page,
-			PerPage: pageSize,
-			Filters: filters}
-		reports, err = d.reporting.ListReports(ctx, query)
-		if err != nil {
-			log.Errorf("Error getting reporting/ListReports %v", err)
-		}
-	}
 }
